Replaces the old TileCompositor with the new one.
