+++ podpac/compositor.py
@@ -5,5 +5,4 @@ Compositor Public Module
 # REMINDER: update api docs (doc/source/user/api.rst) to reflect changes to this file
 
 from podpac.core.compositor.ordered_compositor import OrderedCompositor
-from podpac.core.compositor.tile_compositor import UniformTileCompositor, UniformTileMixin, TileCompositor
-from podpac.core.compositor.data_compositor import DataCompositor
+from podpac.core.compositor.tile_compositor import TileCompositor, TileCompositorRaw
