+++ podpac/core/compositor/test/test_tiled_compositor.py
@@ -4,10 +4,13 @@ import numpy as np
 import podpac
 from podpac.core.data.datasource import DataSource
 from podpac.core.data.array_source import Array, ArrayBase
-from podpac.core.compositor.data_compositor import DataCompositor
+from podpac.core.compositor.tile_compositor import TileCompositorRaw, TileCompositor
 
 
-class TestDataCompositor(object):
+class TestTileCompositor(object):
     def test_basic_composition(self):
         # TODO
+        pass
+
+    def test_interpolated(self):
         pass
