+++ podpac/datalib/test/test_cosmos.py
@@ -11,17 +11,14 @@ class TestCOSMOS(object):
     def test_common_coordinates(self):
         point_interpolation = {
             "method": "nearest",
-            "params": {
-                "use_selector": False,
-                "remove_nan": True,
-                "time_scale": "1,M",
-                "respect_bounds": False,
-            },
+            "params": {"use_selector": False, "remove_nan": True, "time_scale": "1,M", "respect_bounds": False},
         }
         cosmos = podpac.datalib.cosmos_stations.COSMOSStations()
         cosmos_raw = podpac.datalib.cosmos_stations.COSMOSStationsRaw()
         cosmos_filled = podpac.datalib.cosmos_stations.COSMOSStations(interpolation=point_interpolation)
         for ck, c in COORDINATES.items():
+            if ck != "cosmos_region":
+                continue
             print("Evaluating: ", ck)
             o_f = cosmos_filled.eval(c)
             assert np.any(np.isfinite(o_f.data))
